backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
pydantic==2.7.0

neo4j==5.19.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.30

celery==5.4.0
redis==5.0.4

sentence-transformers==3.0.1
numpy==1.26.4
transformers==4.40.0
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.0+cpu

PyPDF2==3.0.1

ollama==0.2.1
httpx==0.27.0

python-dotenv==1.0.1
rich==13.7.1
